requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
psycopg2-binary==2.9.9
django-extensions==3.2.3
